tests/wpt/tests/css/css-fonts/font-face-range-order.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: Order of values in @font-face range descriptors</title>
<link rel="author" title="Dominik Röttsches" href="mailto:drott@chromium.org"/>
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-prop-desc">
<meta name="assert" content="Allow ranges to be specified in reverse order.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<style>
@font-face {
font-family: reversed-range-test;
font-stretch: 200% 50%;
font-style: oblique 90deg -90deg;
font-weight: 900 100;
src: local(Ahem);
}
</style>
<script>
  promise_test(async (t) => {
    const fonts = await document.fonts.load("12px reversed-range-test");
    assert_equals(fonts[0].stretch, "200% 50%", "Stretch value must be returned as specified.");
    assert_equals(fonts[0].style, "oblique 90deg -90deg", "Style value must be returned as specified.");
    assert_equals(fonts[0].weight, "900 100", "Weight value must be returned as specified.");
  });
</script>
</body>
</html>
