LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/font-face-range-order-expected.txt


FAIL CSS Fonts Module Level 3: Order of values in @font-face range descriptors assert_equals: Stretch value must be returned as specified. expected "200% 50%" but got "normal"
